navbar removed, default window setup changed to a pathbar, keyboard shortcuts implemented
