+++ ui/src/components/WindowContainer.tsx
@@ -7,7 +7,7 @@ import useWindowStore from '../state/useWindowStore'
 export default function WindowContainer({
   map,
   id,
-  isVertical
+  isVertical,
 }: WindowContainerProps): JSX.Element {
   const { delWindow } = useWindowStore()
   const lastChange = useRef<number[]>([])
@@ -35,7 +35,7 @@ export default function WindowContainer({
         sizes.length > 1 &&
         JSON.stringify(sizes) != JSON.stringify(lastChange.current)
       ) {
-        const index = sizes.findIndex((num) => num === 0)
+        const index = sizes.findIndex(num => num === 0)
 
         if (index !== -1 && hasChildren) {
           if (index === 0 && map.has(childId)) {
@@ -52,13 +52,6 @@ export default function WindowContainer({
     [hasChildren, childId]
   )
 
-  console.log(
-    id,
-    `has ${childId}`,
-    map.has(childId),
-    `has ${childId + 1}`,
-    map.has(childId + 1)
-  )
 
   if (!map) return <></>
